# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rochelle Ataide Confeitaria ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: available
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── CUSTOMERS ────────────────────────────────────────────────────────────
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: customerId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    patch:
      operationId: updateOrder
      tags: [orders]
      summary: Update order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
    delete:
      operationId: deleteOrder
      tags: [orders]
      summary: Delete order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── PRODUCTION ───────────────────────────────────────────────────────────
  /production:
    get:
      operationId: listProductionOrders
      tags: [production]
      summary: List production orders (today's panel)
      parameters:
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Production orders list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductionOrder"
    post:
      operationId: createProductionOrder
      tags: [production]
      summary: Create a production order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionOrder"
  /production/{id}/status:
    patch:
      operationId: updateProductionStatus
      tags: [production]
      summary: Update production order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductionStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductionOrder"

  # ─── STOCK ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, price, available, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        price:
          type: number
        cost:
          type: ["number", "null"]
        imageUrl:
          type: ["string", "null"]
        available:
          type: boolean
        unit:
          type: ["string", "null"]
        minStock:
          type: ["integer", "null"]
        cmvPercent:
          type: ["number", "null"]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        cost:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        unit:
          type: string
        minStock:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
        price:
          type: number
        cost:
          type: number
        imageUrl:
          type: string
        available:
          type: boolean
        unit:
          type: string
        minStock:
          type: integer

    # ─── CUSTOMER ───────────────────────────────────────────────────────────
...
    Order:
      type: object
      required: [id, status, total, deliveryType, deliveryDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: ["integer", "null"]
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, awaiting_payment, paid, production, ready, out_for_delivery, delivered, cancelled]
        total:
          type: number
        deliveryType:
          type: string
          enum: [pickup, delivery]
        deliveryAddress:
          type: ["string", "null"]
        deliveryDate:
          type: string
        deliveryTime:
          type: ["string", "null"]
        deliveryFee:
          type: number
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, subtotal]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        subtotal:
          type: number
        notes:
          type: ["string", "null"]
    OrderItemInput:
      type: object
      required: [productId, quantity, unitPrice]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        notes:
          type: string
    OrderInput:
      type: object
      required: [deliveryType, deliveryDate, items]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryType:
          type: string
          enum: [pickup, delivery]
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
    OrderUpdate:
      type: object
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        deliveryType:
          type: string
          enum: [pickup, delivery]
        deliveryAddress:
          type: string
        deliveryDate:
          type: string
        deliveryTime:
          type: string
        deliveryFee:
          type: number
        notes:
          type: string
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, awaiting_payment, paid, production, ready, out_for_delivery, delivered, cancelled]

    # ─── PRODUCTION ─────────────────────────────────────────────────────────
    ProductionOrder:
      type: object
      required: [id, productId, productName, quantity, scheduledDate, status, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        scheduledDate:
          type: string
        scheduledTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, producing, finished]
        orderId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    ProductionOrderInput:
      type: object
      required: [productId, quantity, scheduledDate]
      properties:
        productId:
          type: integer
        quantity:
          type: integer
        scheduledDate:
          type: string
        scheduledTime:
          type: string
        orderId:
          type: integer
        notes:
          type: string
    ProductionStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, producing, finished]

    # ─── STOCK ──────────────────────────────────────────────────────────────
...