# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rochelle Ataide Confeitaria ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  categoryId?: number | null;
  /** @nullable */
  categoryName?: string | null;
  price: number;
  /** @nullable */
  cost?: number | null;
  /** @nullable */
  imageUrl?: string | null;
  available: boolean;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  minStock?: number | null;
  /** @nullable */
  cmvPercent?: number | null;
  createdAt: string;
}