# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rochelle Ataide Confeitaria ERP API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description?: string;
  categoryId?: number;
  price: number;
  cost?: number;
  imageUrl?: string;
  available?: boolean;
  unit?: string;
  minStock?: number;
}