# lib/api-zod/src/generated/types/productionOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Rochelle Ataide Confeitaria ERP API
 * OpenAPI spec version: 0.1.0
 */
import type { ProductionOrderStatus } from './productionOrderStatus';

export interface ProductionOrder {
  id: number;
  productId: number;
  productName: string;
  quantity: number;
  scheduledDate: string;
  /** @nullable */
  scheduledTime?: string | null;
  status: ProductionOrderStatus;
  /** @nullable */
  orderId?: number | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}